backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
python-docx==1.1.2
pywin32==308
python-multipart==0.0.20
